% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBaseCommercialDvmt.R
\name{CalculateBaseCommercialDvmt}
\alias{CalculateBaseCommercialDvmt}
\title{Calculate the base year commercial DVMT and ratios with prospective
predictors.}
\usage{
CalculateBaseCommercialDvmt(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateBaseCommercialDvmt} calculate the base year commercial
light-duty vehicle and heavy-duty vehicle DVMT and calculate ratios with
prospective predictor variables.
}
\details{
This function calculates the commercial light-duty DVMT and heavy-duty DVMT
for the base year. It also calculates ratios between DVMT and population,
household DVMT, household income, and workers.
}
