% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFutureCommercialDvmt.R
\name{CalculateFutureCommercialDvmt}
\alias{CalculateFutureCommercialDvmt}
\title{Calculate the future year commercial DVMT and ratios with prospective
predictors.}
\usage{
CalculateFutureCommercialDvmt(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateFutureCommercialDvmt} calculate the future year commercial
light-duty vehicle and heavy-duty vehicle DVMT.
}
\details{
This function calculates the commercial light-duty DVMT and heavy-duty DVMT
for the future year. It may be used to calculate commercial DVMT for any
non-base year.
}
