% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFutureCommercialDvmt.R
\docType{data}
\name{CalculateFutureCommercialDvmtSpecifications}
\alias{CalculateFutureCommercialDvmtSpecifications}
\title{Specifications list for CalculateFutureCommercialDvmt module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{input data specifications}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CalculateFutureCommercialDvmt.R script.
}
\usage{
CalculateFutureCommercialDvmtSpecifications
}
\description{
A list containing specifications for the CalculateFutureCommercialDvmt module.
}
\keyword{datasets}
