% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitLDComDvmt.R
\docType{data}
\name{LDVehAgeCDF_AgTy}
\alias{LDVehAgeCDF_AgTy}
\title{Vehicle age table}
\format{A list having the following components:
\describe{
  \item{LDVehAgeCDF_AgTy}{a table of vehicle cumulative age proportions by vehicle type}
}}
\source{
SplitLDComDvmt.R script.
}
\usage{
LDVehAgeCDF_AgTy
}
\description{
A table of light-duty vehicle age proportions by vehicle type.
}
\keyword{datasets}
