% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitLDComDvmt.R
\docType{data}
\name{SplitLDComDvmtSpecifications}
\alias{SplitLDComDvmtSpecifications}
\title{Specifications list for SplitLDComDvmt module}
\format{A list containing 6 components:
\describe{
 \item{NewInpTable}{table to be created}
 \item{NewSetTable}{table to be created}
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{model inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
SplitLDComDvmt.R script.
}
\usage{
SplitLDComDvmtSpecifications
}
\description{
A list containing specifications for the SplitLDComDvmt module.
}
\keyword{datasets}
